pyproject.toml
[project]
name = "ejemplos-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.73.1",
    "easyocr>=1.7.2",
    "flagembedding>=1.3.5",
    "flashrank>=0.2.10",
    "groq>=0.37.1",
    "hf-xet>=1.2.0",
    "langchain>=1.2.6",
    "langchain-core>=1.2.6",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-qdrant>=1.1.0",
    "python-dotenv>=1.2.1",
    "qdrant-client[fastembed]>=1.16.2",
    "ragas>=0.4.3",
    "semchunk>=2.2.2",
    "sentence-transformers>=5.2.0",
]
